+++ src/__init__.py
@@ -1,0 +1,1 @@
+"""Neko Idle Quest source package."""