+++ src/scenes/battle_scene.py
@@ -1,0 +1,7 @@
+"""Battle scene placeholder."""
+
+from src.scenes.base_scene import BaseScene
+
+
+class BattleScene(BaseScene):
+    """Represent the future auto battle screen."""