+++ src/scenes/menu_scene.py
@@ -1,0 +1,7 @@
+"""Main menu scene placeholder."""
+
+from src.scenes.base_scene import BaseScene
+
+
+class MenuScene(BaseScene):
+    """Represent the future main menu."""